Rename format util file to date_util
