+++ lib/screens/person/person_detail_screen.dart
@@ -13,7 +13,7 @@ import 'package:payment_calendar/widgets/radio_option_tile.dart';
 import '../../models/expense.dart';
 import '../../models/person.dart';
 import '../../providers/expenses_provider.dart';
-import '../../utils/format.dart';
+import '../../utils/date_util.dart';
 import '../expense/expense_detail_screen.dart';
 
 enum _DateFilter { all, thisMonth, lastMonth, custom }
